+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
  - Input on first launch of bot to request Discord bot token
  - Setprefix to have a default parameter value of `?`
  - License from GNU GPLv3 to GNU APGLv3
+ - Balance to have an exact and simplified command
 
 ## [1.0] - 2020-12-31
 ### Added